# AIM-OpenAPI.yaml
openapi: 3.0.3
...
info:
  title: Armory Inventory Management APIs
  description: APIs for managing AIM (Armory Inventory Management)
  version: 1.0.0
...
paths:
...
  /organizations:
    get:
      summary: Get all organizations
      operationId: getAllOrganizations
      tags:
        - Organizations
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
#        - $ref: '#/components/parameters/Authorization'
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: A list of Organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationDto'
#        '500':
#          description: Internal server error

        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"


    post:
      summary: Create a new organization
      operationId: createOrganization
      tags:
        - Organizations
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
#        - $ref: '#/components/parameters/Authorization'
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationDto'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '400':
          description: Bad request - Missing or invalid data
        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
        '409':
          description: Organization with same name already present.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /organizations/{organizationId}:
    get:
      summary: Get organization by organizationId
      operationId: getOrganizationById
      tags:
        - Organizations
      parameters:
        - name: organizationId
          in: path
          required: true
          description: ID of the organization to fetch
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
#        - $ref: '#/components/parameters/Authorization'
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Organization details by organizationId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Organization not found"
        '500':
          description: Internal server error

    put:
      summary: Update organization by organization Id
      operationId: updateOrganization
      tags:
        - Organizations
      parameters:
        - name: organizationId
          in: path
          required: true
          description: ID of the organization to update
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
#        - $ref: '#/components/parameters/Authorization'
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationDto'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
        '404':
          description: Organization data not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Organization not found"
        '400':
          description: Bad request - Missing or invalid data
        '500':
          description: Internal server error

    delete:
      summary: Delete Organization by organizationId
      operationId: deleteOrganization
      tags:
        - Organizations
      parameters:
        - name: organizationId
          in: path
          required: true
          description: ID of the organization to delete
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
#        - $ref: '#/components/parameters/Authorization'
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Organization deleted successfully
        '403':
          description: Forbidden - Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Organization not found"
        '500':
          description: Internal server error

  #=======================================Department APIs===========================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    X-Tenant-ID:
      name: X-Tenant-ID
      in: header
      description: Tenant identifier for multi-tenancy.
      required: true
      schema:
        type: string
#    Authorization:
#      name: Authorization
#      in: header
#      description: "Bearer token for authentication. Format: `Bearer <token>`"
#      required: true
#      schema:
#        type: string
  schemas:
...
    OrganizationDto:
      type: object
      properties:
        organizationId:
          type: integer
          format: int64
          description: The unique identifier of the organization.
        organizationName:
          type: string
          description: The unique Name of the organization.
        organizationDetails:
          type: object
          additionalProperties: true  # Allows storing arbitrary JSON data
          description: A JSON object containing details such as organization name, username, email, and phone number
          example:
            email: "admin@weaponvault.com"
            phone_no: "+1234567890"
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the organization was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the organization was last updated.
        status:
          type: string
          description: The status of the organization (active, inactive, etc.)
#        departments: # One-to-many relationship to departments
#          type: array
#          items:
#            $ref: '#/components/schemas/DepartmentDto'
      required:
        - organizationName
        - organizationDetails

    #Department Schema
    DepartmentDto:
      type: object
      properties:
        departmentId:
          type: integer
          format: int64
          description: The unique identifier of the department.
        departmentName:
          type: string
          description: The name of the department.
        departmentDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing details about the department.
          example:
            department_code: "HR001"
            department_head: "John Doe"
            description: "Handles all human resources tasks."
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the department was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the department was last updated.
#        organization:  # Many-to-one relationship back to the organization
#          $ref: '#/components/schemas/OrganizationDto'
      required:
        - departmentName
        - departmentDetails


    # Manufacturer Schemas
...