# src/main/java/com/armory/WeaponVault/organizationApi/OrganizationsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.armory.WeaponVault.organizationApi;

import com.armory.WeaponVault.entityDTO.GetAllOrganizations403Response;
import com.armory.WeaponVault.entityDTO.GetAllOrganizations500Response;
import com.armory.WeaponVault.entityDTO.GetOrganizationById404Response;
import com.armory.WeaponVault.entityDTO.OrganizationDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-04-29T14:12:51.910033071+05:30[Asia/Kolkata]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "Organizations", description = "Operations related to managing organizations.")
public interface OrganizationsApi {

    default OrganizationsApiDelegate getDelegate() {
        return new OrganizationsApiDelegate() {};
    }

    /**
     * POST /organizations : Create a new organization
     *
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @param organizationDto  (required)
     * @return Organization created successfully (status code 201)
     *         or Bad request - Missing or invalid data (status code 400)
     *         or Forbidden - Access Denied (status code 403)
     *         or Organization with same name already present. (status code 409)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createOrganization",
        summary = "Create a new organization",
        tags = { "Organizations" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Organization created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrganizationDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request - Missing or invalid data"),
            @ApiResponse(responseCode = "403", description = "Forbidden - Access Denied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations403Response.class))
            }),
            @ApiResponse(responseCode = "409", description = "Organization with same name already present."),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations500Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/organizations",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<OrganizationDto> createOrganization(
        @NotNull @Parameter(name = "X-Tenant-ID", description = "Tenant identifier for multi-tenancy.", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = true) String xTenantID,
        @Parameter(name = "OrganizationDto", description = "", required = true) @Valid @RequestBody OrganizationDto organizationDto
    ) {
        return getDelegate().createOrganization(xTenantID, organizationDto);
    }


    /**
     * DELETE /organizations/{organizationId} : Delete Organization by organizationId
     *
     * @param organizationId ID of the organization to delete (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return Organization deleted successfully (status code 200)
     *         or Forbidden - Access Denied (status code 403)
     *         or Organization not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteOrganization",
        summary = "Delete Organization by organizationId",
        tags = { "Organizations" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Organization deleted successfully"),
            @ApiResponse(responseCode = "403", description = "Forbidden - Access Denied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations403Response.class))
            }),
            @ApiResponse(responseCode = "404", description = "Organization not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetOrganizationById404Response.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/organizations/{organizationId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteOrganization(
        @Parameter(name = "organizationId", description = "ID of the organization to delete", required = true, in = ParameterIn.PATH) @PathVariable("organizationId") Long organizationId,
        @NotNull @Parameter(name = "X-Tenant-ID", description = "Tenant identifier for multi-tenancy.", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = true) String xTenantID
    ) {
        return getDelegate().deleteOrganization(organizationId, xTenantID);
    }


    /**
     * GET /organizations : Get all organizations
     *
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return A list of Organizations (status code 200)
     *         or Forbidden - Access Denied (status code 403)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getAllOrganizations",
        summary = "Get all organizations",
        tags = { "Organizations" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of Organizations", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OrganizationDto.class)))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden - Access Denied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations403Response.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations500Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/organizations",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<OrganizationDto>> getAllOrganizations(
        @NotNull @Parameter(name = "X-Tenant-ID", description = "Tenant identifier for multi-tenancy.", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = true) String xTenantID
    ) {
        return getDelegate().getAllOrganizations(xTenantID);
    }


    /**
     * GET /organizations/{organizationId} : Get organization by organizationId
     *
     * @param organizationId ID of the organization to fetch (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return Organization details by organizationId (status code 200)
     *         or Forbidden - Access Denied (status code 403)
     *         or Organization not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getOrganizationById",
        summary = "Get organization by organizationId",
        tags = { "Organizations" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Organization details by organizationId", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrganizationDto.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden - Access Denied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations403Response.class))
            }),
            @ApiResponse(responseCode = "404", description = "Organization not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetOrganizationById404Response.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/organizations/{organizationId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<OrganizationDto> getOrganizationById(
        @Parameter(name = "organizationId", description = "ID of the organization to fetch", required = true, in = ParameterIn.PATH) @PathVariable("organizationId") Long organizationId,
        @NotNull @Parameter(name = "X-Tenant-ID", description = "Tenant identifier for multi-tenancy.", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = true) String xTenantID
    ) {
        return getDelegate().getOrganizationById(organizationId, xTenantID);
    }


    /**
     * PUT /organizations/{organizationId} : Update organization by organization Id
     *
     * @param organizationId ID of the organization to update (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @param organizationDto  (required)
     * @return Organization updated successfully (status code 200)
     *         or Forbidden - Access Denied (status code 403)
     *         or Organization data not found (status code 404)
     *         or Bad request - Missing or invalid data (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateOrganization",
        summary = "Update organization by organization Id",
        tags = { "Organizations" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Organization updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrganizationDto.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden - Access Denied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllOrganizations403Response.class))
            }),
            @ApiResponse(responseCode = "404", description = "Organization data not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetOrganizationById404Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request - Missing or invalid data"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/organizations/{organizationId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<OrganizationDto> updateOrganization(
        @Parameter(name = "organizationId", description = "ID of the organization to update", required = true, in = ParameterIn.PATH) @PathVariable("organizationId") Long organizationId,
        @NotNull @Parameter(name = "X-Tenant-ID", description = "Tenant identifier for multi-tenancy.", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = true) String xTenantID,
        @Parameter(name = "OrganizationDto", description = "", required = true) @Valid @RequestBody OrganizationDto organizationDto
    ) {
        return getDelegate().updateOrganization(organizationId, xTenantID, organizationDto);
    }

}